refactor: save satellite state in dictionaries and simplify track
Dictionaries.jl provides an efficient way to save satellite state
in dictionaries. It is a lot easier to merge / remove satellite state
with this commit while at the same time it is still efficient.
Moreover, state that's needed for doppler estimation or to
downconvert and correlate the signal is now directly saved into
SatState, which removes the difficulty to sync with the corresponding
hierarchic of satellite system -> satellite prn.