feat: use bit detector buffer to fill bit buffer once found
The integrated code blocks used to detect the secondary code or
bit were neglected in the bit buffer.
This has been changed with this commit.
Once the bit has been found all already integrated code blocks
will be used to fill the bit buffer.
In some rare cases this can lead to faster decoding of the
satellite frames.