feat: simplify cpu buffers (#62)
I think it was a misconception to have the temporary buffers
inside the tracking state. Strictly speaking they were only for
internal use and it wasn't a state that was going to be used in the
next iteration. Hence, it made more sense to pass these buffers
directly to the `track` function.
There was, however, one drawback of this approach: I didn't want the
user to care about these buffers and, hence, there needed to be
a default value, which have been passed to the track function.
Having a default value, however, would have meant that the
buffers would have been allocated each time the track function
had been called. While search for a solution for this, I came
across Bumper.jl which has some neat features, that alleviates
the need to allocate buffers beforehand without performance
compromise. That's what I have done here.